backend/pyproject.toml
[project]
name = "vtuber-live-backend"
version = "1.0.0"
description = "VTuber Live Aggregator Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.40",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.4",
    "alembic>=1.13.0",
    "yt-chat-downloader>=0.0.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "yt-dlp>=2024.8.6",
    "aiosqlite>=0.22.1",
    "loguru>=0.7.0",
    "bilibili-api-python>=17.4.1",
    "curl-cffi>=0.14.0",
    "slowapi>=0.1.9",
    "beautifulsoup4>=4.12.0",
    "cloudscraper>=1.2.71",
    "redis>=7.4.0",
    "tenacity>=9.0.0",
    "tqdm>=4.67.3",
    "rich>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
]

[dependency-groups]
dev = [
    "httpx",
    "pytest",
]
